requirements.txt
google-genai
openai
pyyaml
python-dotenv
feedparser
httpx
beautifulsoup4
streamlit
plotly
pandas
python-dateutil
docx2txt
openpyxl
python-pptx
win10toast; platform_system == "Windows"
